1.Main activity: add button 2.Entry manage activity : button save , update main act 3.Main act: switch List menu 4.Main act: add switch 5.switch initial data 6.toggle button color

+++ app/src/main/java/com/rmc/thienle/jedi2/implementation/services/RelayServiceImpl.java
@@ -89,7 +89,7 @@ public class RelayServiceImpl extends DataBaseHelper implements RelayService {
     }
 
     @Override
-    public List<Relay> getRelayBySwitchId(int switchId) {
+    public List<Relay> getRelayBySwitchId(long switchId) {
         Cursor res =  db.rawQuery("select * from "+ RelayImpl.RelayDetails.TABLE_NAME + " where " + RelayImpl.RelayDetails.COLUMN_NAME_SWITCH_ID + "=" + switchId, null );
         if(res.getCount()>0){
             List<Relay> relaylist = new ArrayList<>();
